LF-5418 Request the dashboard ticket through RTK Query
The endpoint moves to store/api/dashboardTicketApi.ts as a mutation, where
new endpoints belong, and prepareHeaders supplies the Authorization header.
handOffToDashboardIfRequested becomes a plain async function in
containers/dashboardTicketHandoff.ts; yield call accepts it unchanged, so
the three sign-in sagas keep the same two lines and the watcher and its
action are no longer needed.

request.reset() drops the mutation result once the ticket has been read.
The whole store is persisted to localStorage, so without it the ticket
would be written to disk.

Starting the endpoint needs a cast to the thunk dispatch signature, because
store.ts annotates its middleware array as Middleware[] and that reduces
store.dispatch to Dispatch<AnyAction>.
