+++ packages/webapp/src/containers/CustomSignUp/saga.js
@@ -25,7 +25,7 @@ import { axios } from '../saga';
 import { enqueueErrorSnackbar } from '../Snackbar/snackbarSlice';
 import { getLanguageFromLocalStorage } from '../../util/getLanguageFromLocalStorage';
 import { setCustomSignUpErrorKey, setPasswordResetError } from '../customSignUpSlice';
-import { handOffToDashboardIfRequested } from '../dashboardTicketSaga';
+import { handOffToDashboardIfRequested } from '../dashboardTicketHandoff';
 
 const loginUrl = (email) => `${url}/login/user/${email}`;
 const loginWithPasswordUrl = () => `${url}/login`;
