+++ packages/webapp/src/main.jsx
@@ -24,7 +24,6 @@ import homeSaga from './containers/saga';
 import addFarmSaga from './containers/AddFarm/saga';
 import peopleSaga from './containers/Profile/People/saga';
 import signUpSaga from './containers/CustomSignUp/saga';
-import dashboardTicketSaga from './containers/dashboardTicketSaga';
 import resetUserPasswordSaga from './containers/PasswordResetAccount/saga';
 import outroSaga from './containers/Outro/saga';
 import documentSaga from './containers/Documents/saga';
@@ -82,7 +81,6 @@ sagaMiddleware.run(homeSaga);
 sagaMiddleware.run(addFarmSaga);
 sagaMiddleware.run(peopleSaga);
 sagaMiddleware.run(signUpSaga);
-sagaMiddleware.run(dashboardTicketSaga);
 sagaMiddleware.run(resetUserPasswordSaga);
 sagaMiddleware.run(outroSaga);
 sagaMiddleware.run(managementPlanSaga);
